chore(#29): startVoiceSession has 2x less code
I decided to opt for using generators to loop through the queue. So long as the generator yields the next audio resource, the bot will keep playing after idle until no more resources are yielded. This simplification means that in /quickplay, the generator will only ever yield once so it will disconnect on first idle. However in /play, the generator will infinitely yield using the while(true) {} loop until no more audio resources can be yielded. In addition to this, the VOICE_CONNECTION_SIGNALS constant has been removed as a nice side-effect to this simplification!